Update to v2.1
Host of minor changes:

Schedules can be selected via drop down menu, will then calculate the
new schedule for the day. The available options are prepopulated by
files in the schedule directory. If they have 'database' in the name
they are dynamic.

User comment now makes the text box the focus

User comments now flagged with ** instead of ## with an additional ** at
the end to make them stand out more

Philosophical change to schedule calculation. Daily information is
calculated and held in a seperate class, the schedule calcualtion then
uses this class to calculate the schedule, the daily information is
calculated at startup and at reset. This speeds up the procedure if you
are changing schedules during the day.